+++ src/builder/commands/generic.rs
@@ -4,7 +4,7 @@ use unshare::{Command};
 
 use super::super::context::Context;
 use super::super::super::path_util::ToRelative;
-use process_util::{capture_stdout, run_and_wait};
+use process_util::{capture_stdout};
 use builder::error::StepError;
 
 
@@ -127,7 +127,7 @@ pub fn command<P:AsRef<Path>>(ctx: &Context, cmdname: P)
 pub fn run(mut cmd: Command) -> Result<(), StepError> {
     debug!("Running {:?}", cmd);
 
-    match run_and_wait(&mut cmd) {
+    match cmd.status() {
         Ok(ref s) if s.success() => Ok(()),
         Ok(s) => Err(StepError::CommandFailed(Box::new(cmd), s)),
         Err(e) => Err(StepError::CommandError(Box::new(cmd), e)),